netplay fix

+++ src/ConsoleNetplay/NetplayRuntimeSupport.h
@@ -225,7 +225,7 @@ struct RuntimeRollbackProcessState
     FrameNumber lastMissingRollbackSnapshotFrame = 0;
     FrameNumber lastMissingRollbackSnapshotLocalFrame = 0;
     FrameNumber lastRecoveryReanchorFrame = 0;
-    uint8_t consecutiveMissingRollbackSnapshotDeferrals = 0;
+    uint32_t consecutiveMissingRollbackSnapshotDeferrals = 0;
 };
 
 struct RuntimeRollbackProcessSettings
